+++ poptorch/source/PoptorchSymbols.cpp
@@ -155,10 +155,6 @@ c10::Symbol optimizer_group;
 c10::Symbol begin_multi_conv;
 c10::Symbol multi_conv_part;
 c10::Symbol end_multi_conv;
-c10::Symbol begin_autocast;
-c10::Symbol suppress_autocast;
-c10::Symbol restore_autocast;
-c10::Symbol autocast;
 
 c10::Symbol update_param_inplace;
 
@@ -200,11 +196,7 @@ static void initializePoptorchSymbols() {
   SYMBOL_INIT(poptorch, begin_multi_conv);
   SYMBOL_INIT(poptorch, multi_conv_part);
   SYMBOL_INIT(poptorch, end_multi_conv);
-  SYMBOL_INIT(poptorch, begin_autocast);
-  SYMBOL_INIT(poptorch, suppress_autocast);
-  SYMBOL_INIT(poptorch, restore_autocast);
   SYMBOL_INIT(poptorch, host_side_cast);
-  SYMBOL_INIT(poptorch, autocast);
 
   SYMBOL_INIT(poptorch, update_param_inplace);