+++ poptorch/source/dispatch_tracer/RegisterAtenOverloads.cpp
@@ -936,19 +936,6 @@ void identityLoss(const c10::OperatorHandle &op, c10::Stack *stack) {
   }
 }
 
-void autocastOp(const c10::OperatorHandle &op, c10::Stack *stack) {
-  ERROR_ON_MSG(
-      poptorch::isDispatcherOn(),
-      "The autocast API is not supported in PopTorch while using the "
-      "dispatcher frontend (the default since version 3.0); normal PyTorch "
-      "casting should be used to replicate its behaviour. For more information "
-      "on porting autocast code, see "
-      "https://docs.graphcore.ai/projects/poptorch-user-guide/en/latest/"
-      "supported_ops.html#bit-float-migration");
-
-  opWithNoReturn(op, stack);
-}
-
 // TODO(T64770) This method is the old way of registering custom functions. The
 // new way would look like this:
 //
@@ -1047,15 +1034,6 @@ static auto registry =
                 .schema("poptorch::pop_name_scope() -> ()")
                 .catchAllKernel<PTC(opWithNoReturn)>())
         .op(torch::RegisterOperators::options()
-                .schema("poptorch::begin_autocast() -> ()")
-                .catchAllKernel<PTC(autocastOp)>())
-        .op(torch::RegisterOperators::options()
-                .schema("poptorch::suppress_autocast() -> ()")
-                .catchAllKernel<PTC(autocastOp)>())
-        .op(torch::RegisterOperators::options()
-                .schema("poptorch::restore_autocast() -> ()")
-                .catchAllKernel<PTC(autocastOp)>())
-        .op(torch::RegisterOperators::options()
                 .schema("poptorch::end_cpu_op(Tensor[] output) -> Tensor[]")
                 .catchAllKernel<PTC(endCpuOp)>())
         .op(torch::RegisterOperators::options()
@@ -1111,10 +1089,6 @@ TORCH_LIBRARY_IMPL(poptorch, AutogradXLA, m) {
   m.impl("end_multi_conv", torch::autograd::autogradNotImplementedFallback());
   m.impl("push_name_scope", torch::autograd::autogradNotImplementedFallback());
   m.impl("pop_name_scope", torch::autograd::autogradNotImplementedFallback());
-  m.impl("begin_autocast", torch::autograd::autogradNotImplementedFallback());
-  m.impl("suppress_autocast",
-         torch::autograd::autogradNotImplementedFallback());
-  m.impl("restore_autocast", torch::autograd::autogradNotImplementedFallback());
   m.impl("end_cpu_op", torch::autograd::autogradNotImplementedFallback());
   m.impl("call_cpu_op", torch::autograd::autogradNotImplementedFallback());
   m.impl("set_attribute", torch::autograd::autogradNotImplementedFallback());