+++ docs/user_guide/changelog.rst
@@ -20,6 +20,8 @@ API changes
 - In addition to tensors, :py:func:`~poptorch.set_overlap_for_input` and :py:func:`~poptorch.set_overlap_for_output` can now be applied to tuples, lists, and dicts of tensors.
 - `poptorch.Options.useOfflineIpuTarget` accepts ``ipu_version = 21`` (Mk2x ipu).
 - ``POPTORCH_IPU_MODEL_VERSION`` accepts ``ipu21`` (Mk2x ipu).
+- The mixed-precision automatic casting API, `autocast`, has now been removed.
+  Instead, please use PyTorch's built-in casting facility.
 
 
 Bug Fixes